Initial commit
This project provides a mechanism to specify types for class properties.
Also, property types have additional parameters, such as the regular
expression pattern for strings, minimum and maximum value for integers
etc. The special setter method handles type checking and will set the
given value only if it is valid for the given type and its additional
parameters. Check future documentation for an example.