Create type classes for integer and floating point numbers
The integer type and the floating point type support the minimum and
the maximum value specification. The setter method will not set the
value if it is outside of the bounds of the minimum and the maximum
value defined in the class property annotation. Initial values are the
minimum and the maximum values defined as constants in PHP for the
integer and the floating point type.