Create an example to show how this mechanism could be used
The Student class has protected properties (it is important that they
are protected and not private) which have annotation comments with the
@type tag set to the fully-qualified class name (path includes the name
space) and any additional parameters that the particular type might
support. Setter methods are made so that they use the setProperty
method of the base class for all typed parameter classes. It will
handle value checking and setting if it meats the required
specifications defined by the type and any additional parameter values.
Here, in this implementation, setters return the self instance ($this)
so that multiple setters can be used in a chain. This is not a
requirement. It is done so here out of convenience.